backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
ragas==0.1.21
langchain>=0.2.5
langchain-anthropic>=0.1.20
langchain-community>=0.2.5
anthropic>=0.31.0
pydantic>=2.7.0
python-dotenv>=1.0.0
datasets>=2.20.0
pandas>=2.2.0
httpx>=0.27.0
sse-starlette>=1.8.0
numpy>=1.26.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
